# swagger.json
{
	"components": {
...
		"schemas": {
...
			"IUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"uri": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"tokens": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"projects": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"username",
					"uri",
					"email",
					"tokens",
					"projects"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReturnUser": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUser"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegisterRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "lbd-middleware",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnUser"
								},
								"examples": {
									"Example 1": {
										"value": {
											"user": {
												"username": "max",
												"uri": "https://lbdserver.org/max",
												"email": "max@mustermann.be",
												"projects": [],
												"tokens": []
											},
											"token": "theBearerToken"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Register as a local user on the LBDserver. The body of the request contains a \"username\" field (used to create a WebID that can be used for Access Control purposes), an \"email\" field and a \"password\".",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterRequest"
							}
						}
					}
				}
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnUser"
								},
								"examples": {
									"Example 1": {
										"value": {
											"user": {
												"username": "max",
												"uri": "https://lbdserver.org/max",
												"email": "max@mustermann.be",
												"projects": [],
												"tokens": []
											},
											"token": "theBearerToken"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/logout": {
			"post": {
				"operationId": "Login",
				"responses": {
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}